+++ src/kong_svelte/package.json
@@ -29,18 +29,18 @@
     "@dicebear/collection": "^9.2.2",
     "@dicebear/core": "^9.2.2",
     "@dicebear/identicon": "9.2.2",
-        "@windoge98/plug-n-play": "0.0.61-siws-8",
-
     "bignumber.js": "9.1.2",
-    "chart.js": "4.5.0",
+    "chart.js": "4.4.8",
     "chartjs-adapter-date-fns": "3.0.0",
+    "comlink": "4.4.2",
     "d3": "7.9.0",
     "date-fns": "4.1.0",
     "emoji-picker-element": "^1.26.1",
     "html5-qrcode": "2.3.8",
     "hyperid": "3.3.0",
+    "localforage": "^1.10.0",
     "qrcode": "1.5.4",
-    "three": "0.177.0"
+    "three": "0.176.0"
   },
   "devDependencies": {
     "@sveltejs/adapter-static": "3.0.8",
@@ -51,28 +51,28 @@
     "@types/node": "^22.13.4",
     "@types/pouchdb-core": "7.0.15",
     "@types/qrcode": "1.5.5",
-    "@types/three": "0.177.0",
+    "@windoge98/plug-n-play": "0.0.61-siws-8",
     "autoprefixer": "10.4.20",
     "dotenv": "16.4.7",
-    "eslint": "^9.20.1",
+    "eslint": "9.20.1",
     "eslint-plugin-svelte": "3.0.2",
-    "jsdom": "26.1.0",
+    "jsdom": "26.0.0",
     "lodash-es": "4.17.21",
-    "lucide-svelte": "0.518.0",
-    "postcss": "8.5.6",
-    "prettier": "3.5.3",
-    "prettier-eslint": "16.4.2",
-    "sass": "1.89.2",
-    "svelte": "5.34.7",
-    "svelte-language-server": "0.17.16",
+    "lucide-svelte": "0.486.0",
+    "postcss": "8.5.3",
+    "prettier": "3.5.1",
+    "prettier-eslint": "16.3.0",
+    "sass": "1.85.0",
+    "svelte": "5.25.6",
+    "svelte-language-server": "0.17.10",
     "svelte-portal": "2.2.1",
     "tailwindcss": "3.4.14",
-    "terser": "5.43.1",
-    "typescript": "5.8.3",
-    "vite": "6.3.5",
+    "terser": "5.39.0",
+    "typescript": "5.8.2",
+    "vite": "6.1.1",
     "vite-plugin-compression": "0.5.1",
     "vite-plugin-environment": "1.1.3",
-    "vite-plugin-pwa": "1.0.0",
-    "vitest": "3.2.4"
+    "vite-plugin-pwa": "0.21.1",
+    "vitest": "3.0.6"
   }
 }
