--- src/kong_svelte/src/lib/types/index.ts
@@ -1,4 +1,0 @@
-// Common types used across the application
-
-// Period type for leaderboard and other time-based data
-export type Period = 'day' | 'week' | 'month';
