--- src/kong_svelte/src/routes/(app)/predict/tos/+page.svelte
@@ -1,145 +1,0 @@
-<script lang="ts">
-  // You can add any necessary script logic here if needed
-</script>
-
-<div class="prose prose-invert max-w-none text-kong-text-secondary">
-  <h1 class="text-3xl font-bold mb-6 text-kong-text-primary">Terms of Service</h1>
-
-  <p class="mb-4">Last updated: April 22, 2025</p>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">1. Introduction</h2>
-  <p class="mb-4">
-    Welcome to Kong Prediction Markets. These Terms of Service ("Terms") govern your access to and use
-    of our decentralized prediction market platform and related services (collectively, the "Service"). 
-    By accessing or using the Service, you agree to be bound by these Terms. These Terms constitute a 
-    legally binding agreement between you and Kong Prediction Markets.
-  </p>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">2. Eligibility</h2>
-  <p class="mb-4">
-    You must be at least 18 years old and capable of forming a binding contract to use the Service.
-    You represent and warrant that you meet these requirements and that your use of the Service does not 
-    violate any laws or regulations in your jurisdiction. You agree not to access or use the Service if 
-    you are located in, under the control of, or a national or resident of any country to which 
-    Switzerland has embargoed goods or services.
-  </p>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">3. Your Account</h2>
-  <p class="mb-4">
-    You are responsible for maintaining the confidentiality of your account credentials, private keys,
-    and wallet information and for all activities that occur under your account. You agree to notify us 
-    immediately of any unauthorized use of your account. Since our platform operates on decentralized 
-    technology, we cannot recover lost private keys or passwords. You are solely responsible for 
-    safeguarding your cryptographic assets.
-  </p>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">4. Use of the Service</h2>
-  <p class="mb-4">
-    You agree to use the Service only for lawful purposes and in accordance with these Terms. You are
-    solely responsible for your conduct and any data, text, information, usernames, graphics, images,
-    photos, profiles, audio and video clips, links ("Content") that you submit, post, and display on
-    the Service.
-  </p>
-  <p class="mb-4">
-    Prohibited activities include, but are not limited to:
-  </p>
-  <ul class="list-disc ml-8 mb-4">
-    <li>Market manipulation or any attempt to manipulate prediction outcomes</li>
-    <li>Creating or participating in markets related to illegal activities</li>
-    <li>Using the platform for money laundering or financing illicit activities</li>
-    <li>Attempting to interfere with the platform's functionality or security</li>
-    <li>Creating markets designed to harm individuals or promote violence</li>
-    <li>Impersonating other users or Kong team members</li>
-    <li>Attempting to circumvent any technological measures implemented for legal compliance</li>
-  </ul>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">5. Decentralized Nature of the Service</h2>
-  <p class="mb-4">
-    You understand and acknowledge that Kong Prediction Markets operates on blockchain technology and 
-    functions as a decentralized application. This means:
-  </p>
-  <ul class="list-disc ml-8 mb-4">
-    <li>Transactions are irreversible and cannot be modified once confirmed on the blockchain</li>
-    <li>Market resolutions are determined by smart contracts according to predefined rules</li>
-    <li>Your interactions with the platform are recorded on a public blockchain</li>
-    <li>Delays, congestion, or fluctuations in the underlying blockchain may affect your experience</li>
-    <li>We have limited ability to recover lost funds or reverse transactions</li>
-  </ul>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">6. Risks</h2>
-  <p class="mb-4">
-    Participating in prediction markets involves significant risks, including but not limited to:
-  </p>
-  <ul class="list-disc ml-8 mb-4">
-    <li>Financial risk: Market prices can be volatile, and you may lose your entire investment</li>
-    <li>Smart contract risks: Despite security audits, smart contracts may contain bugs or vulnerabilities</li>
-    <li>Regulatory risks: Changes in regulations may affect the operation of the platform or specific markets</li>
-    <li>Oracle risks: Market resolutions depend on oracles that provide external data</li>
-    <li>Cryptocurrency volatility: The value of cryptocurrencies used for transactions may fluctuate</li>
-  </ul>
-  <p class="mb-4">
-    By using the Service, you explicitly acknowledge and accept these risks.
-  </p>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">7. Disclaimers</h2>
-  <p class="mb-4">
-    THE SERVICE IS PROVIDED "AS IS" AND "AS AVAILABLE" WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS
-    OR IMPLIED, INCLUDING, BUT NOT LIMITED TO, IMPLIED WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
-    PARTICULAR PURPOSE, AND NON-INFRINGEMENT. WE DO NOT WARRANT THAT THE SERVICE WILL BE UNINTERRUPTED,
-    ERROR-FREE, OR SECURE. KONG DOES NOT GUARANTEE THE ACCURACY OF MARKET RESOLUTIONS, ORACLE DATA, 
-    OR THE BEHAVIOR OF SMART CONTRACTS.
-  </p>
-  <p class="mb-4">
-    NO ADVICE OR INFORMATION, WHETHER ORAL OR WRITTEN, OBTAINED FROM KONG OR THROUGH THE SERVICE, WILL 
-    CREATE ANY WARRANTY NOT EXPRESSLY MADE HEREIN. THE SERVICE IS NOT INTENDED TO PROVIDE FINANCIAL, 
-    LEGAL, OR INVESTMENT ADVICE.
-  </p>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">8. Limitation of Liability</h2>
-  <p class="mb-4">
-    TO THE FULLEST EXTENT PERMITTED BY LAW, IN NO EVENT WILL KONG OR ITS AFFILIATES BE LIABLE FOR ANY
-    INDIRECT, INCIDENTAL, SPECIAL, CONSEQUENTIAL, OR PUNITIVE DAMAGES, OR ANY LOSS OF PROFITS OR
-    REVENUES, WHETHER INCURRED DIRECTLY OR INDIRECTLY, OR ANY LOSS OF DATA, USE, GOODWILL, OR OTHER
-    INTANGIBLE LOSSES, RESULTING FROM (a) YOUR ACCESS TO OR USE OF OR INABILITY TO ACCESS OR USE THE
-    SERVICE; (b) ANY CONDUCT OR CONTENT OF ANY THIRD PARTY ON THE SERVICE; (c) UNAUTHORIZED ACCESS,
-    USE, OR ALTERATION OF YOUR TRANSMISSIONS OR CONTENT; (d) SMART CONTRACT FAILURES OR VULNERABILITIES;
-    (e) ORACLE MALFUNCTIONS OR INACCURACIES; OR (f) BLOCKCHAIN NETWORK CONGESTION OR FAILURES.
-  </p>
-  <p class="mb-4">
-    IN NO EVENT SHALL KONG'S TOTAL LIABILITY TO YOU FOR ALL CLAIMS EXCEED THE AMOUNT YOU HAVE PAID 
-    TO KONG FOR ACCESS TO THE SERVICE IN THE PAST SIX MONTHS.
-  </p>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">9. Compliance with Laws</h2>
-  <p class="mb-4">
-    You are responsible for complying with all applicable laws and regulations in your jurisdiction, 
-    including tax obligations related to profits or losses from prediction markets. The Service is not 
-    offered to citizens or residents of jurisdictions where participation in prediction markets is 
-    prohibited. You understand that regulatory frameworks for prediction markets and cryptocurrency 
-    are evolving, and changes may impact your ability to use the Service. Kong Prediction Markets 
-    operates in compliance with Swiss regulations regarding virtual assets and financial services.
-  </p>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">10. Governing Law</h2>
-  <p class="mb-4">
-    These Terms shall be governed by and construed in accordance with the laws of Switzerland, 
-    without regard to its conflict of law principles. Any dispute arising from or relating to the subject 
-    matter of these Terms shall be finally settled in Zürich, Switzerland, in accordance with Swiss law.
-    You agree that the courts of Switzerland shall have exclusive jurisdiction to resolve any dispute
-    arising out of or relating to these Terms.
-  </p>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">11. Changes to Terms</h2>
-  <p class="mb-4">
-    We reserve the right to modify these Terms at any time. We will provide notice of material changes
-    by posting the updated Terms on the Service or through other communications. Your continued use of
-    the Service after such notice constitutes your acceptance of the modified Terms. If you do not agree 
-    to the modified Terms, you must stop using the Service.
-  </p>
-
-  <h2 class="text-xl font-semibold mt-6 mb-2 text-kong-text-primary">12. Contact Us</h2>
-  <p class="mb-4">
-    If you have any questions about these Terms, please contact us at support@kongprediction.com or 
-    via our Discord community at t.me/kong_swap.
-  </p>
-</div>
